chore: refactor package/shared to have constants/ and /types separately
