+++ apps/server/src/controller-server/ControllerContext.ts
@@ -2,7 +2,7 @@
  * @license
  * Copyright 2025 BrowserOS
  */
-import { TIMEOUTS } from '@browseros/shared/timeouts'
+import { TIMEOUTS } from '@browseros/shared/constants/timeouts'
 
 import type { Context } from '../tools/controller-based/index.js'
