+++ apps/server/src/main.ts
@@ -8,7 +8,7 @@
 
 import fs from 'node:fs'
 import path from 'node:path'
-import { RATE_LIMITS } from '@browseros/shared/limits'
+import { RATE_LIMITS } from '@browseros/shared/constants/limits'
 import { RateLimiter } from './agent/index.js'
 import {
   ensureBrowserConnected,
